88: Rename tests to match test subject package names
